rename: remove select stories
